LayoutTests/imported/w3c/web-platform-tests/fetch/range/non-matching-range-response-expected.txt


PASS Range requests with no rewrites should succeed
PASS Range response out of range of request should succeed
FAIL Range requests ignored (200 status) should succeed assert_equals: expected "ok" but got "timeout"
